ai-chatbot/requirements.txt
fastapi
uvicorn
pydantic==2.*
openai
chromadb
tiktoken
python-dotenv
pymysql
